Pulled BucketWriter to a separate class
