+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/api/functions/sink/filesystem/Bucket.java
@@ -56,7 +56,7 @@ public class Bucket<IN, BucketID> {
 
 	private final int subtaskIndex;
 
-	private final InProgressFileWriter.BucketWriter<IN, BucketID> bucketWriter;
+	private final BucketWriter<IN, BucketID> bucketWriter;
 
 	private final RollingPolicy<IN, BucketID> rollingPolicy;
 
@@ -81,7 +81,7 @@ public class Bucket<IN, BucketID> {
 			final BucketID bucketId,
 			final Path bucketPath,
 			final long initialPartCounter,
-			final InProgressFileWriter.BucketWriter<IN, BucketID> partFileFactory,
+			final BucketWriter<IN, BucketID> partFileFactory,
 			final RollingPolicy<IN, BucketID> rollingPolicy,
 			final OutputFileConfig outputFileConfig) {
 		this.subtaskIndex = subtaskIndex;
@@ -104,7 +104,7 @@ public class Bucket<IN, BucketID> {
 	private Bucket(
 			final int subtaskIndex,
 			final long initialPartCounter,
-			final InProgressFileWriter.BucketWriter<IN, BucketID> partFileFactory,
+			final BucketWriter<IN, BucketID> partFileFactory,
 			final RollingPolicy<IN, BucketID> rollingPolicy,
 			final BucketState<BucketID> bucketState,
 			final OutputFileConfig outputFileConfig) throws IOException {
@@ -349,7 +349,7 @@ public class Bucket<IN, BucketID> {
 	 * @param bucketId the identifier of the bucket, as returned by the {@link BucketAssigner}.
 	 * @param bucketPath the path to where the part files for the bucket will be written to.
 	 * @param initialPartCounter the initial counter for the part files of the bucket.
-	 * @param partFileFactory the {@link InProgressFileWriter.BucketWriter} the factory creating part file writers.
+	 * @param partFileFactory the {@link BucketWriter} the factory creating part file writers.
 	 * @param <IN> the type of input elements to the sink.
 	 * @param <BucketID> the type of the identifier of the bucket, as returned by the {@link BucketAssigner}
 	 * @param outputFileConfig the part file configuration.
@@ -360,7 +360,7 @@ public class Bucket<IN, BucketID> {
 			final BucketID bucketId,
 			final Path bucketPath,
 			final long initialPartCounter,
-			final InProgressFileWriter.BucketWriter<IN, BucketID> partFileFactory,
+			final BucketWriter<IN, BucketID> partFileFactory,
 			final RollingPolicy<IN, BucketID> rollingPolicy,
 			final OutputFileConfig outputFileConfig) {
 		return new Bucket<>(subtaskIndex, bucketId, bucketPath, initialPartCounter, partFileFactory, rollingPolicy, outputFileConfig);
@@ -370,7 +370,7 @@ public class Bucket<IN, BucketID> {
 	 * Restores a {@code Bucket} from the state included in the provided {@link BucketState}.
 	 * @param subtaskIndex the index of the subtask creating the bucket.
 	 * @param initialPartCounter the initial counter for the part files of the bucket.
-	 * @param partFileFactory the {@link InProgressFileWriter.BucketWriter} the factory creating part file writers.
+	 * @param partFileFactory the {@link BucketWriter} the factory creating part file writers.
 	 * @param bucketState the initial state of the restored bucket.
 	 * @param <IN> the type of input elements to the sink.
 	 * @param <BucketID> the type of the identifier of the bucket, as returned by the {@link BucketAssigner}
@@ -380,7 +380,7 @@ public class Bucket<IN, BucketID> {
 	static <IN, BucketID> Bucket<IN, BucketID> restore(
 			final int subtaskIndex,
 			final long initialPartCounter,
-			final InProgressFileWriter.BucketWriter<IN, BucketID> partFileFactory,
+			final BucketWriter<IN, BucketID> partFileFactory,
 			final RollingPolicy<IN, BucketID> rollingPolicy,
 			final BucketState<BucketID> bucketState,
 			final OutputFileConfig outputFileConfig) throws IOException {
